Chapter03/03_05/end/catalog-service/src/test/java/com/polarbookshop/catalogservice/domain/BookServiceTest.java
package com.polarbookshop.catalogservice.domain;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.time.Year;

import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class BookServiceTest {

    @Mock
    private BookRepository bookRepository;

    @InjectMocks
    private BookService bookService;

    @Test
    void whenBookToCreateAlreadyExistsThenThrows() {
        String bookIsbn = "1234561232";
        Book bookToCreate = new Book(bookIsbn, "Title", "Author", Year.of(2000));
        when(bookRepository.existsByIsbn(bookIsbn)).thenReturn(true);
        assertThatThrownBy(() -> bookService.addBookToCatalog(bookToCreate))
                .isInstanceOf(BookAlreadyExistsException.class)
                .hasMessage("A book with ISBN " + bookIsbn + " already exists.");
    }

    @Test
    void whenBookToUpdateNotFoundThenThrows() {
        String bookIsbn = "1234561234";
        Book bookToUpdate = new Book(bookIsbn, "Title", "Author", Year.of(2000));
        when(bookRepository.existsByIsbn(bookIsbn)).thenReturn(false);
        assertThatThrownBy(() -> bookService.editBookDetails(bookIsbn, bookToUpdate))
                .isInstanceOf(BookNotFoundException.class)
                .hasMessage("The book with ISBN " + bookIsbn + " was not found.");
    }
}
